pyproject.toml
[project]
name = "drillapi"
version = "0.1.2"
description = "drillapi"
authors = [{name = "SFOE", email = "geoinformation@bfe.admin.ch"}]
readme = "README.md"
requires-python = ">=3.14"
keywords = ["drillapi"]
dependencies = [
  "fastapi>=0.128.0",
  "httpx>=0.28.1",
  "jinja2>=3.1.6",
  "mangum>=0.21.0",
  "owslib>=0.35.0",
  "pydantic-settings>=2.12.0",
  "slowapi>=0.1.9",
  "uvicorn>=0.40.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "requests",
  "respx",
  "pytest-asyncio",
  "ruff>=0.14.14",
]

[project.scripts]
drillapi = "drillapi.__main__:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
